requirements.txt
# requirements.txt
ccxt
pandas
numpy
scikit-learn
python-telegram-bot
APScheduler
python-dotenv
Flask
gunicorn